backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pandas
sqlalchemy